backend:refactor split auth views
